erf dgt grgr Signed-off-by: yuxs <yuxs2022@163.com>

+++ have.c
@@ -1,5 +1,5 @@
 #include <stdio.h>
- float main()
+ int main()
  {
     /*变量定义*/
     /*<类型名称》《变量名称》
@@ -8,19 +8,19 @@
        int price ,amount 
              c的保留字
                 auto,break,case,cher,const,continue,default */
-    float price = 0;/*定义一个变量，名字是price ，类型是int，初始值是0*/
+    int price = 0;/*定义一个变量，名字是price ，类型是int，初始值是0*/
  
     printf("请输入商品金额（元）");
 
     scanf("%d", &price);
 
-    float amount = 0;
+    int amount = 0;
 
     printf("请输入给定的总金额（元）");
 
     scanf("%d", &amount);
 
-    float change = amount - price;
+    int change = amount - price;
 
     printf("找你%d元。 \n",change);
